[DEV-6368] Move invalidate_precomputed_on_labels_change() to miners package
