+++ server/athenian/api/internal/miners/github/deployment.py
@@ -6,7 +6,17 @@ from itertools import chain, groupby, product, repeat
 import json
 import logging
 from operator import attrgetter
-from typing import Any, Collection, Iterator, KeysView, Mapping, NamedTuple, Optional, Sequence
+from typing import (
+    Any,
+    AsyncIterator,
+    Collection,
+    Iterator,
+    KeysView,
+    Mapping,
+    NamedTuple,
+    Optional,
+    Sequence,
+)
 
 import aiomcache
 import medvedi as md
@@ -3468,3 +3478,80 @@ async def reset_broken_deployments(account: int, pdb: Database, rdb: Database) -
             len(rows_commits),
         )
     return len(bad_deps)
+
+
+async def invalidate_precomputed_on_labels_change(
+    deployment_name: str,
+    changed_labels: Sequence[str],
+    account: int,
+    prefixer: Prefixer,
+    logical_settings: LogicalRepositorySettings,
+    rdb: Database,
+    pdb: Database,
+) -> None:
+    """Invalidate the precomputed facts when the labels associated with a deployment change.
+
+    If the physical repositories referred by the changed deployment has any logical
+    repositories, and their deployment settings refers any of the changed labels,
+    then all deployments notified after this one referring involved repositories are invalidated.
+
+    """
+    log = logging.getLogger(f"{metadata.__package__}.invalidate_precomputed_on_labels_change")
+
+    async def _discover_affected_repos() -> AsyncIterator[int]:
+        changed_labels_set = set(changed_labels)
+
+        repo_ids_stmt = sa.select(DeployedComponent.repository_node_id).where(
+            DeployedComponent.account_id == account,
+            DeployedComponent.deployment_name == deployment_name,
+        )
+        repo_id_rows = await rdb.fetch_all(repo_ids_stmt)
+        repo_ids = [r[DeployedComponent.repository_node_id.name] for r in repo_id_rows]
+
+        for repo_id in repo_ids:
+            repo_name = prefixer.repo_node_to_name[repo_id]
+            try:
+                logical_depl_settings = logical_settings.deployments(repo_name)
+            except KeyError:  # repos pointed by deployment has no logical repos
+                continue
+
+            # check if its logical repos use any of the changed labels
+            # TODO: avoid access to private LogicalDeploymentSettings._labels
+            if changed_labels_set.intersection(logical_depl_settings._labels):
+                yield repo_id
+
+    affected_repos = [repo_id async for repo_id in _discover_affected_repos()]
+    if not affected_repos:
+        log.info("Labels %s change doesn't required precomputed invalidation", changed_labels)
+        return
+
+    finished_at_subselect = (
+        sa.select(DeploymentNotification.finished_at)
+        .where(
+            DeploymentNotification.account_id == account,
+            DeploymentNotification.name == deployment_name,
+        )
+        .scalar_subquery()
+    )
+    select_from = sa.join(
+        DeploymentNotification,
+        DeployedComponent,
+        onclause=sa.and_(
+            DeploymentNotification.account_id == DeployedComponent.account_id,
+            DeploymentNotification.name == DeployedComponent.deployment_name,
+        ),
+    )
+    where = [
+        DeploymentNotification.account_id == account,
+        DeployedComponent.repository_node_id.in_(affected_repos),
+        DeploymentNotification.finished_at > finished_at_subselect,
+    ]
+
+    names_stmt = sa.select(DeploymentNotification.name).select_from(select_from).where(*where)
+    name_rows = await rdb.fetch_all(names_stmt)
+    names = [r[DeploymentNotification.name.name] for r in name_rows]
+
+    to_delete = names + [deployment_name]
+    log.info("Invalidating precomputed %s for labels %s change", to_delete, changed_labels)
+
+    await _delete_precomputed_deployments(names + [deployment_name], account, pdb)
